// src/modules/inpatient/inform-consent/models/approval-or-refusal-of-anesthetic-action-model.ts
export interface IApprovalOrRefusalOfAnestheticActionFormModel {
  ID_Dokter_Pelaksana: string;
  Nama_Dokter_Pelaksana: string;
  ID_Pemberi_Informasi: string;
  Nama_Pemberi_Informasi: string;
  Penerima_Informasi: string;
  Diagnosis_Check: string;
  Dasar_Diagnosis_Check: string;
  Dasar_Klinis: string;
  Dasar_Radiologi: string;
  Dasar_EKG: string;
  Dasar_Laboratorium: string;
  Tindakan_Kedokteran_Check: string;
  Anestesi_Umum_Intubasi_Check: string;
  Anestesi_Umum_LMA_Check: string;
  Anestesi_Umum_FM_Check: string;
  Anestesi_Umum_TIVA_Check: string;
  Anestesi_Regional_Spinal_Check: string;
  Anestesi_Regional_Epidural_Check: string;
  Anestesi_Regional_Perifer_Check: string;
  Indikasi_Tindakan_Tujuan_Check: string;
  Tata_Cara_Tindakan_Check: string;
  Tata_Cara_Tindakan: string;
  Risiko_Check: string;
  Shock_Check: string;
  Henti_Jantung_Check: string;
  Meninggal_Check: string;
  Tujuan_Check: string;
  Tujuan: string;
  Komplikasi_Check: string;
  Komplikasi_Umum_1: string;
  Komplikasi_Umum_2: string;
  Komplikasi_Umum_3: string;
  Komplikasi_Umum_4: string;
  Komplikasi_Umum_5: string;
  Komplikasi_Umum_6: string;
  Komplikasi_Umum_7: string;
  Komplikasi_Umum_8: string;
  Komplikasi_Umum_9: string;
  Komplikasi_Umum_10: string;
  Komplikasi_Regional_1: string;
  Komplikasi_Regional_2: string;
  Komplikasi_Regional_3: string;
  Komplikasi_Regional_4: string;
  Komplikasi_Regional_5: string;
  Komplikasi_Regional_6: string;
  Komplikasi_Regional_7: string;
  Komplikasi_Regional_8: string;
  Komplikasi_Regional_9: string;
  Komplikasi_Regional_10: string;
  Komplikasi_Regional_11: string;
  Komplikasi_Regional_12: string;
  Prognosis_Check: string;
  Prognosis: string;
  Alternatif_Tindakan_Check: string;
  Alternatif_Tindakan: string;
  Lain_Lain_Check: string;
  Lain_Lain: string;
  ID_Dokter_Pelaksana_TTD: string;
  TTD_Dokter_Pelaksana: string;
  Nama_Dokter_Pelaksana_TTD: string;
  TTD_Penerima_Informasi: string;
  Waktu: string;
  Kota: string;
  Pernyataan_Id: string;
  Tanda_Tangan: string;
  Tanda_Tangan_Pasien: string;
  Tanda_Tangan_Saksi: string;
  ID_Saksi: string;
  Nama_Saksi: string;
  Pasien_Kota: string;
  Pasien_Tanggal: string;
  Updated_At: string;
  Updated_By: string;
  Updated_By_Name: string;
  Tanda_Tangan_Nama: string;
  Tanda_Tangan_Tgl_Lahir: string;
  Tanda_Tangan_JK: string;
  Tanda_Tangan_Telp: string;
  Tanda_Tangan_Alamat: string;
  Tindakan_Operasi: string;
  ID_Petugas: string;
  Nama_Petugas: string;
  Nama_Saksi_Keluarga: string;
  Tanda_Tangan_Saksi_2: string;
  Tanda_Tangan_Hubungan: string;
}
// ...
export interface IApprovalOrRefusalOfAnestheticActionModel extends IDataModel {
  form: IApprovalOrRefusalOfAnestheticActionFormModel;
}

export class ApprovalOrRefusalOfAnestheticActionModel extends DataModel {
  form?: IApprovalOrRefusalOfAnestheticActionFormModel;
  rawat_inap: any;
  constructor(request: IApprovalOrRefusalOfAnestheticActionModel) {
    super(request);
    // if (request.form) {
    this.form = request.form;
    // }
  }
}